backend/requirements.txt
flask
flask-cors
gunicorn

# Deep learning frameworks
torch
torchvision
tensorflow-cpu

# Computer vision
opencv-python-headless
numpy
Pillow
mediapipe

# Utilities
gdown